Updated functions to accept any length integer.

+++ BitwiseExamples/setResetBit.cpp
@@ -3,8 +3,8 @@
 
 using namespace std;
 
-//sets the bit (=1) at the provided position (0-7) in the provided 8 bit integer
-int setBit(uint8_t num, int bitPos) {
+//sets the bit (=1) at the provided position in the provided 8 bit integer
+int setBit(int num, int bitPos) {
     //takes 8 bit num [ b7 b6 b5 b4 b3 b2 b1 b0 ] and sets the bit to "1" by ORing with a shifted 0b1. Ex, bitPos = 4 [ 0 0 0 1 0 0 0 0]
     // Ex: [ b7 b6 b5 b4 b3 b2 b1 b0 ]
     //  OR [  0  0  0  1  0  0  0  0 ]
@@ -13,8 +13,8 @@ int setBit(uint8_t num, int bitPos) {
     return num;
 }
 
-//resets the bit (=0) at the provided position (0-7) in the provided 8 bit integer
-int resetBit(uint8_t num, int bitPos) {
+//resets the bit (=0) at the provided position in the provided 8 bit integer
+int resetBit(int num, int bitPos) {
     //takes 8 bit num [ b7 b6 b5 b4 b3 b2 b1 b0 ] and sets the bit to "0" by ANDing with the inverse of a shifted 0b1. Ex, bitPos = 4 [ 1 1 1 0 1 1 1 1]
     // Ex: [ b7 b6 b5 b4 b3 b2 b1 b0 ]
     // AND [  1  1  1  0  1  1  1  1 ]     -->   ~[ 0 0 0 1 0 0 0 0 ] = [ 1 1 1 0 1 1 1 1 ]
